Revert "drag and drop"
