+++ js/components/DataAndFunction/DataPanel.js
@@ -1,0 +1,20 @@
+import React from "react";
+import DataBlock  from "./DataBlock";
+
+
+
+
+export default class DataPanel extends React.Component {
+    
+    render() {
+const { articles } = this.props
+
+        return (
+ <div class="data-panel">
+            
+        <DataBlock  articles = { articles }> </DataBlock>
+ </div>
+            
+      );
+  }
+}
