--- src/css/style.css
@@ -1,350 +1,0 @@
-
-html, body {
-  height: 100%;
-  cursor: default !important;
-}
-
-li {
-  list-style: none;
-}
-
-.leftside{
-  position:absolute;
-  left:0;
-  vertical-align: middle;
-  top:50%;
- cursor:pointer;
-}
-
-.rightside{
-  cursor:pointer;
-  position:absolute;
-  right:0;
-  vertical-align: middle;
-  top:50%;
-
-
-}
-
-#content {
-  height: 3000px;
-  background: #649173;
-  background: -webkit-linear-gradient(30deg, #373B44 , #649173);
-  background: linear-gradient(30deg, #373B44 , #649173);
-  -webkit-transition: background 3s;
-  transition: background 3s;
-}
-
-.content-0 {
-  background: #135058 !important;
-  background: -webkit-linear-gradient(30deg, #649173 , #135058) !important;
-  background: linear-gradient(30deg, #649173 , #135058) !important;
-}
-
-.content-1 {
-  background: #996666 !important;
-  background: -webkit-linear-gradient(30deg, #CC9933 , #996666) !important;
-  background: linear-gradient(30deg, #CC9933 , #996666) !important;
-}
-
-.content-2 {
-  background: #666666 !important;
-  background: -webkit-linear-gradient(30deg, #D39D38 , #666666) !important;
-  background: linear-gradient(30deg, #D39D38 , #666666) !important;
-}
-
-#wrapper {
-  height: 100%;
-  padding: 20px;
-  -webkit-transition: filter .3s, -webkit-filter .3s;
-  transition: filter .3s, -webkit-filter .3s;
-}
-
-#loadingOverlay {
-  display: table;
-  position: fixed;
-  background-color: rgba(0, 0, 0, .1);
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 10000;
-}
-
-.blur-content {
-  -webkit-filter: blur(2px);
-  filter: blur(2px);
-  -webkit-filter: progid\:DXImageTransform\.Microsoft\.Blur(PixelRadius\=2,MakeShadow\=false);
-  filter: progid\:DXImageTransform\.Microsoft\.Blur(PixelRadius\=2,MakeShadow\=false);
-}
-
-.data-panel {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 240px;
-  padding: 10px;
-  font-size: 20px;
-  z-index: 9999;
-}
-
-.function-panel {
-  position: fixed;
-  top: 0;
-  right: 0;
-  height: 100%;
-  width: 130px;
-  padding: 10px;
-  font-size: 20px;
-  z-index: 9999;
-}
-
-.display-panel {
-  margin-left: 240px;
-  margin-right: 130px;
-  width: 1000px;
-  height: 2960px;
-  font-size: 20px;
-  -webkit-transition: background .3s, opacity .3s, filter .3s, -webkit-filter .3s;
-  transition: background .3s, opacity .3s, filter .3s, -webkit-filter .3s;
-  z-index: 0;
-}
-.detail-panel {
-  width:500px;
-  height:250px;
-}
-
-.help-bg {
-  background-image: url(../img/help-bg.svg);
-  background-repeat: no-repeat;
-  background-size: 100%;
-}
-
-.drop-target {
-  opacity: .7;
-  border-radius: 30px;
-  -webkit-filter: blur(1px);
-  filter: blur(1px);
-  -webkit-filter: progid\:DXImageTransform\.Microsoft\.Blur(PixelRadius\=1,MakeShadow\=false);
-  filter: progid\:DXImageTransform\.Microsoft\.Blur(PixelRadius\=1,MakeShadow\=false);
-}
-
-.function-button {
-  display: block;
-  margin: 10px auto;
-  width: 100%;
-  cursor: pointer;
-}
-
-.switch-button {
-  margin-top: 30px;
-  width: 100%;
-  font-size: 14px;
-  color: #fff;
-}
-
-.switch-button:hover {
-  color: #eee;
-}
-
-.switch-button-overlay .ant-popover-inner-content {
-  padding: 0;
-}
-
-.switch-button-overlay ul, .switch-button-overlay li {
-  border: none;
-}
-
-.switch-button-overlay li {
-  line-height: 36px;
-  height: 36px;
-  padding-left: 15px;
-}
-
-.switch-button-overlay li>span {
-  cursor: pointer;
-  transition: color .2s;
-  display: inline-block;
-}
-
-.switch-button-overlay li>i {
-  float: right;
-  line-height: 36px;
-  margin-right: 10px;
-  margin-left: 10px;
-}
-
-.switch-button-overlay li>span:hover {
-  color: #57c5f7;
-}
-
-.switch-button-overlay li[disabled]>span {
-  color: #ddd;
-  cursor: not-allowed;
-}
-
-.data-card, .line-card, .pie-card {
-  position: absolute;
-  box-shadow: 1px 2px 10px rgba(22, 22, 22, 0.5);
-  z-index: 100;
-  transition: border-radius .3s, box-shadow .3s, filter .3s, -webkit-filter .3s;
-}
-
-.data-card:hover, .line-card:hover, .pie-card:hover {
-  position: absolute;
-  box-shadow: 1px 2px 20px rgba(22, 22, 22, 0.7);
-}
-
-.data-card {
-  width: auto;
-  cursor: default;
-}
-
-.line-card {
-  width: auto;
-  cursor: default;
-}
-
-.line-card .line {
-  width: 500px;
-  height: 300px;
-}
-
-.pie-card {
-  width: auto;
-}
-
-.pie-card .ant-card-body {
-  display: table;
-}
-
-.pie-card .pie {
-  display: table-cell;
-  height: 300px;
-  width: 350px;
-  vertical-align: middle;
-}
-
-.pie-card .inf-detail {
-  display: table-cell;
-  padding: 20px 20px 20px 12px;
-  line-height: 20px;
-  vertical-align: middle;
-}
-
-.pie-card ul {
-  margin-left: 10px;
-}
-
-.tile {
-  display: inline-block;
-  margin: 5px 10px;
-  min-width: 150px;
-  width: auto;
-  cursor: pointer;
-}
-
-.data-block {
-  margin: 5px auto;
-  padding: 5px 10px;
-  cursor: move;
-  border-radius: 8px;
-  -webkit-transition: background-color .2s;
-  transition: background-color .2s;
-}
-
-.data-block:hover {
-  background-color: rgba(255,255,255,0.3);
-}
-
-.data-item, .data-title {
-  display: block;
-}
-
-.data-item {
-  margin: 10px;
-  padding: 3px 10px;
-}
-
-.data-title {
-  color: #fff;
-  font-weight: 600;
-  font-size: 20px;
-  -webkit-transition: color .2s;
-  transition: color .2s;
-  text-shadow: 0 0 3px #666;
-}
-
-.ant-btn {
-  border-color: #fff!important;
-  color: #fff!important;
-  text-shadow: 0 0 3px #666;
-  box-shadow: 0 0 1px #666;
-}
-
-.ant-btn:hover {
-  border-color: #eee;
-  color: #eee;
-}
-
-.anticon-cross {
-  cursor: pointer;
-}
-
-.anticon-cross[disabled] {
-  cursor: not-allowed;
-}
-
-#loadingOverlay > .spinner {
-  display: table-cell;
-  vertical-align: middle;
-  text-align: center;
-  font-size: 10px;
-}
-
-#loadingOverlay > .spinner > div {
-  background-color: #fff;
-  box-shadow: 0 0 3px #666;
-  width: 6px;
-  height: 40px;
-  display: inline-block;
-
-  -webkit-animation: sk-stretchdelay 1.2s infinite ease-in-out;
-  animation: sk-stretchdelay 1.2s infinite ease-in-out;
-}
-
-#loadingOverlay > .spinner .rect2 {
-  -webkit-animation-delay: -1.1s;
-  animation-delay: -1.1s;
-}
-
-#loadingOverlay > .spinner .rect3 {
-  -webkit-animation-delay: -1.0s;
-  animation-delay: -1.0s;
-}
-
-#loadingOverlay > .spinner .rect4 {
-  -webkit-animation-delay: -0.9s;
-  animation-delay: -0.9s;
-}
-
-#loadingOverlay > .spinner .rect5 {
-  -webkit-animation-delay: -0.8s;
-  animation-delay: -0.8s;
-}
-
-@-webkit-keyframes sk-stretchdelay {
-  0%, 40%, 100% { -webkit-transform: scaleY(0.4) }
-  20% { -webkit-transform: scaleY(1.0) }
-}
-
-@keyframes sk-stretchdelay {
-  0%, 40%, 100% {
-    transform: scaleY(0.4);
-    -webkit-transform: scaleY(0.4);
-  }  20% {
-    transform: scaleY(1.0);
-    -webkit-transform: scaleY(1.0);
-  }
-}
