--- src/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>React Tutorials</title>
-  <!-- Bootstrap Core CSS -->
-
-  <!-- Custom Fonts -->
-  
-  </head>
-
-  <body>
-      <div id="content">
-        <div id="app"></div>
-        <div>good</div>
-        <script src="client.min.js"></script>
-        <link href="css/style.css" rel="stylesheet" type="text/css">
-      </div>
-  </body>
-</html>
