--- src/js/components/DataAndFunction/DataPanel.js
@@ -1,20 +1,0 @@
-import React from "react";
-import DataBlock  from "./DataBlock";
-
-
-
-
-export default class DataPanel extends React.Component {
-    
-    render() {
-		const { articles } = this.props
-
-        return (
- 			<div className="data-panel">
-            
-        		<DataBlock  articles = { articles }/> 
- 			</div>
-            
-      );
-  }
-}
