--- src/js/reducers/tweetReducer.js
@@ -1,24 +1,0 @@
-export default function reducer(state={
-    tweets:[],
-    fetching:false,
-    fetched:false,
-    error:null
-    },action){
-    
-    switch(action.type)
-    {
-        case "FETCH_TWEETS":
-        { 
-            return {...state,fetching:true};
-                 
-        }
-        case "FETCH_TWEETS_FULFILLED":
-        {
-    
-            return {...state,fetching:false,fetched:true,tweets:action.payload,}
-        }
-            
-    }
-    
-    return state;
-}
